% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALLCatcheR.R
\name{allcatch_bcrabl1}
\alias{allcatch_bcrabl1}
\title{Classification}
\usage{
allcatch_bcrabl1(Counts.file = NA, ID_class = "symbol", sep = "\\t", out.file = "/path)
}
\arguments{
\item{Counts.file}{count data}

\item{ID_class}{gene ids}

\item{sep}{file seperator}
}

\item{out.file}{output file path}
}
\value{
data.frame containing class predictions
}
\description{
Prediction of B-ALL leukemia subtypes based on expression data
}
\examples{
allcatch_bcrabl1()

}
