% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bcrabl1_models}
\alias{bcrabl1_models}
\title{classification model}
\format{
list with different prediction models
}
\usage{
data(bcrabl1_models)
}
\description{
model used for predictions including cross-validation
}
\keyword{datasets}
